Move transport type fixtures to top level
