+++ raiden/tests/integration/fixtures/transport.py
@@ -1,41 +1,8 @@
-from enum import Enum
-
 import pytest
 
 from raiden.network.transport import MatrixTransport
+from raiden.tests.fixtures.variables import TransportProtocol
 from raiden.tests.utils.transport import generate_synapse_config, matrix_server_starter
-
-
-class TransportProtocol(Enum):
-    UDP = 'udp'
-    MATRIX = 'matrix'
-
-
-@pytest.fixture
-def transport(request):
-    """ 'all' replaced by parametrize in conftest.pytest_generate_tests """
-    return request.config.getoption('transport')
-
-
-@pytest.fixture
-def transport_protocol(transport):
-    return TransportProtocol(transport)
-
-
-@pytest.fixture
-def skip_if_not_udp(request):
-    """Skip the test if not run with UDP transport"""
-    if request.config.option.transport in ('udp', 'all'):
-        return
-    pytest.skip('This test works only with UDP transport')
-
-
-@pytest.fixture
-def skip_if_not_matrix(request):
-    """Skip the test if not run with Matrix transport"""
-    if request.config.option.transport in ('matrix', 'all'):
-        return
-    pytest.skip('This test works only with Matrix transport')
 
 
 @pytest.fixture
